fix mobile view
